dap/functionapps/sftpclient/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.21.3
azure-storage-blob==12.24.1
paramiko==3.5.1
azure-keyvault-secrets==4.9.0
azure-identity==1.20.0
pyzipper==0.3.6
pandas==2.2.3
pyarrow==19.0.1
chardet==5.2.0
pgpy==0.6.0
openpyxl==3.1.5
pyodbc==5.2.0
cryptography==43.0.3
xlrd==2.0.1
azure-storage-queue==12.12.0